Update error handling
